feat: embed whiteboard as a playground activity tab
Extract the whiteboard into a reusable component and open it as a
single-instance playground tab from the activity bar, with canvas-only
pinch zoom that does not steal scroll from overlays like Cloud Sync.
